Make the Separate Intefaces According to the modules
